Make FakeClient a global fixture
