+++ tests/unit/test_cli.py
@@ -3,55 +3,7 @@ import logging
 import pytest
 import yaml
 
-from vault_cli import cli, client, exceptions, settings
-
-
-class FakeClient(client.VaultClientBase):
-    def __init__(self, **kwargs):
-        self.init_kwargs = kwargs
-        self.deleted = []
-        self.db = {}
-
-    def get_secret(self, path):
-        try:
-            return self.db[path]
-        except KeyError:
-            raise exceptions.VaultSecretDoesNotExist(404, "nooooo")
-
-    def list_secrets(self, path):
-        # Just reproducing in memory the behaviour of the real list_secrets
-        # This is complicated enough to have its unit test, below (test_fake_client)
-        paths = [key for key in self.db if key.startswith(path)]
-        result = []
-        for element in paths:
-            element = element[len(path) + 1 if path else 0 :].split("/", 1)
-            if len(element) == 1:
-                result.append(element[0])
-            else:
-                result.append(f"{element[0]}/")
-
-        return sorted(set(result) - {""})
-
-    def _set_secret(self, path, value):
-        self.db[path] = value
-
-    def delete_secret(self, path):
-        self.db.pop(path, None)
-
-
-@pytest.fixture
-def backend(mocker):
-    backend = FakeClient()
-    mocker.patch("vault_cli.requests.RequestsVaultClient", return_value=backend)
-    yield backend
-
-
-@pytest.mark.parametrize(
-    "path, expected", [("", ["a", "b/"]), ("a", []), ("b", ["c", "d"])]
-)
-def test_fake_client_list_secrets(backend, path, expected):
-    backend.db = {"a": "A", "b/c": "BC", "b/d": "BD"}
-    assert backend.list_secrets(path) == expected
+from vault_cli import cli, settings
 
 
 def test_bad_backend(cli_runner, backend):
